requirements.txt
annotated-types>=0.6.0
anyio>=4.3.0
bcrypt>=4.1.2
certifi>=2024.2.2
cffi>=1.16.0
click>=8.1.7
colorama>=0.4.6
contourpy>=1.3.0
cryptography>=42.0.2
cycler>=0.12.1
distlib>=0.3.8
et_xmlfile>=2.0.0
exceptiongroup>=1.2.0
fastapi>=0.110.0
filelock>=3.13.1
Flask>=2.2.2
fonttools>=4.55.0
future>=1.0.0
idna>=3.6
igraph>=0.10.8
importlib_metadata>=8.5.0
importlib_resources>=6.4.5
itsdangerous>=2.2.0
Jinja2>=3.1.4
kiwisolver>=1.4.7
MarkupSafe>=3.0.2
matplotlib>=3.9.2
n2g>=0.3.3
networkx>=3.2.1
numpy>=2.0.2
openpyxl>=3.1.5
packaging>=24.2
paramiko>=3.4.0
pillow>=10.2.0
pipenv>=2023.12.1
platformdirs>=4.2.0
pycparser>=2.21
pydantic>=2.6.3
pydantic_core>=2.16.3
PyNaCl>=1.5.0
pyparsing>=3.2.0
pyserial>=3.5
pyte>=0.8.2
python-dateutil>=2.9.0.post0
python-igraph>=0.10.8
six>=1.16.0
sniffio>=1.3.1
starlette>=0.36.3
sv-ttk>=2.6.0
textfsm>=1.1.3
texttable>=1.7.0
ttp>=0.9.5
ttp-templates>=0.3.7
typing_extensions>=4.10.0
virtualenv>=20.25.1
wcwidth>=0.2.13
Werkzeug>=3.1.3
zipp>=3.21.0
